refactor: Start a new file for tag values tests; move the one existing test
